+++ R/RFPE.R
@@ -2,7 +2,7 @@
 #'
 #' This function computes the robust Final Prediction Errors (RFPE) for a robust regression fit using M-estimates.
 #'
-#' @param object the \code{MM} element (of class \code{\link{lmrob}}) in an object of class \code{\link{lmrobdetMM}}.
+#' @param object an object of class \code{lmrobdetMM} or \code{lmrobM}.
 #' @param scale a numeric value specifying the scale estimate used to compute the RFPE. Usually this 
 #' should be the scale estimate from an encompassing model. If \code{NULL}, the scale estimate in 
 #' \code{object} is used. 
@@ -27,6 +27,9 @@
 #' @export
 lmrobdetMM.RFPE <- function(object, scale = NULL, bothVals = FALSE)
 {
+  if( (class(object)[1] != 'lmrobdetMM') & 
+      (class(object)[1] != 'lmrobM') )
+    stop('RFPE should only be calculated for M or MM regression fits.')
   if (!object$converged)
     warning("The algorithm did not converge, inference is not recommended.")
   # ocm <- tolower(object$control$method)
@@ -355,7 +358,8 @@ step.lmrobdetMM <- function (object, scope, direction = c("both", "backward", "f
     Terms <- terms(update(formula(fit), eval(parse(text = paste("~ .", change)))))
     attr(Terms, "formula") <- new.formula <- formula(Terms)
     # control$method <- 'MM'
-    newfit <- lmrobdetMM(new.formula, data = m, control = control) #, init=object$init$control$method)
+    tmp.call <- call('lmrobdetMM', formula=new.formula, control=control, data = objectcall$data)
+    newfit <- eval.parent(tmp.call) # lmrobdetMM(new.formula, data = m, control = control) #, init=object$init$control$method)
     bRFPE <- aod[, "RFPE"][o]
     if (trace)
       cat("\nStep:  RFPE =", format(round(bRFPE, 4)), "\n",